+++ price-feeder/config/config.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/BurntSushi/toml"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/go-playground/validator/v10"
 )
 
@@ -47,6 +48,10 @@ var (
 		ProviderCoinbase: {},
 		ProviderMock:     {},
 	}
+
+	// maxDeviationThreshold is the maxmimum allowed amount of standard
+	// deviations which validators are able to set for a given asset.
+	maxDeviationThreshold = sdk.MustNewDecFromStr("3.0")
 )
 
 type (
@@ -54,6 +59,7 @@ type (
 	Config struct {
 		Server          Server         `toml:"server"`
 		CurrencyPairs   []CurrencyPair `toml:"currency_pairs" validate:"required,gt=0,dive,required"`
+		Deviations      []Deviation    `toml:"deviation_thresholds"`
 		Account         Account        `toml:"account" validate:"required,gt=0,dive,required"`
 		Keyring         Keyring        `toml:"keyring" validate:"required,gt=0,dive,required"`
 		RPC             RPC            `toml:"rpc" validate:"required,gt=0,dive,required"`
@@ -77,6 +83,13 @@ type (
 		Base      string   `toml:"base" validate:"required"`
 		Quote     string   `toml:"quote" validate:"required"`
 		Providers []string `toml:"providers" validate:"required,gt=0,dive,required"`
+	}
+
+	// Deviation defines a maximum amount of standard deviations that a given asset can
+	// be from the median without being filtered out before voting.
+	Deviation struct {
+		Base      string `toml:"base" validate:"required"`
+		Threshold string `toml:"threshold" validate:"required"`
 	}
 
 	// Account defines account related configuration that is related to the Umee
@@ -124,7 +137,7 @@ type (
 		//
 		// Example:
 		// [["chain_id", "cosmoshub-1"]]
-		GlobalLabels [][]string `toml:"global_labels""`
+		GlobalLabels [][]string `toml:"global_labels"`
 
 		// Type determines which type of telemetry to use
 		// Valid values are "prometheus" or "generic"
@@ -227,6 +240,17 @@ func ParseConfig(configPath string) (Config, error) {
 	}
 	if len(gatePairs) > 1 {
 		return cfg, fmt.Errorf("gate provider does not support multiple pairs: %v", gatePairs)
+	}
+
+	for _, deviation := range cfg.Deviations {
+		threshold, err := sdk.NewDecFromStr(deviation.Threshold)
+		if err != nil {
+			return cfg, fmt.Errorf("deviation thresholds must be numeric: %w", err)
+		}
+
+		if threshold.GT(maxDeviationThreshold) {
+			return cfg, fmt.Errorf("deviation thresholds must not exceed 3.0")
+		}
 	}
 
 	return cfg, cfg.Validate()