+++ CHANGELOG.md
@@ -52,6 +52,19 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - [913](https://github.com/umee-network/umee/pull/913) Changed update registry gov proposal to add and update tokens, but never delete them.
 - [918](https://github.com/umee-network/umee/pull/918) Add MarketSummary query to CLI
 
+### Improvements
+
+- [935](https://github.com/umee-network/umee/pull/935) Fix protobuf linting
+- [959](https://github.com/umee-network/umee/pull/959) Improve ModuleBalance calculation
+- [962](https://github.com/umee-network/umee/pull/962) Streamline AccrueAllInterest
+- [967](https://github.com/umee-network/umee/pull/962) Use taylor series of e^x for more accurate interest at high APY.
+- [987](https://github.com/umee-network/umee/pull/987) Streamline x/leverage CLI tests
+- [1012](https://github.com/umee-network/umee/pull/1012) Improve negative time elapsed error message
+
+### Bug Fixes
+
+- [1018](https://github.com/umee-network/umee/pull/1018) Return nil if negative time elapsed from the last block happens.
+
 ## [v2.0.2](https://github.com/umee-network/umee/releases/tag/v2.0.2) - 2022-05-13
 
 ### Features